backend/requirements.txt
google-cloud-firestore
pydantic